+++ lib/commanded/process_managers/process_manager_instance.ex
@@ -166,7 +166,7 @@ defmodule Commanded.ProcessManagers.ProcessManagerInstance do
     %RecordedEvent{correlation_id: correlation_id, event_id: event_id, event_number: event_number} =
       event
 
-    %State{application: application, idle_timeout: idle_timeout} = state
+    %State{idle_timeout: idle_timeout} = state
 
     case handle_event(event, state) do
       {:error, error} ->
@@ -182,7 +182,7 @@ defmodule Commanded.ProcessManagers.ProcessManagerInstance do
 
       commands ->
         # Copy event id, as causation id, and correlation id from handled event.
-        opts = [application: application, causation_id: event_id, correlation_id: correlation_id]
+        opts = [causation_id: event_id, correlation_id: correlation_id, returning: false]
 
         with :ok <- commands |> List.wrap() |> dispatch_commands(opts, state, event) do
           process_state = mutate_state(event, state)
@@ -301,10 +301,6 @@ defmodule Commanded.ProcessManagers.ProcessManagerInstance do
 
     case application.dispatch(command, opts) do
       :ok ->
-        dispatch_commands(pending_commands, opts, state, last_event)
-
-      # when include_execution_result is set to true, the dispatcher returns an :ok tuple
-      {:ok, _} ->
         dispatch_commands(pending_commands, opts, state, last_event)
 
       error ->