+++ test/aggregates/aggregate_concurrency_test.exs
@@ -96,10 +96,8 @@ defmodule Commanded.Aggregates.AggregateConcurrencyTest do
              }
     end
 
-    test "should error after too many attempts", context do
-      %{account_number: account_number} = context
-
-      # fail to append to stream
+    test "should error after too many attempts", %{account_number: account_number} do
+      # Fail to append to stream
       expect(MockEventStore, :append_to_stream, 6, fn _event_store_meta,
                                                       ^account_number,
                                                       1,