+++ test/commands/dispatch_consistency_test.exs
@@ -19,7 +19,6 @@ defmodule Commanded.Commands.DispatchConsistencyTest do
 
   setup do
     start_supervised!(ConsistencyApp)
-
     :ok
   end